Refactor chatboxes on Teachers page
Solves #70
